requirements.txt
customtkinter
sqlalchemy
apscheduler
plyer
pystray
Pillow